+++ mathics/builtin/lists.py
@@ -268,20 +268,6 @@ class Delete(Builtin):
         return newexpr
 
 
-class Failure(Builtin):
-    """
-    <url>:WMA link:https://reference.wolfram.com/language/ref/Failure.html</url>
-
-    <dl>
-      <dt>Failure[$tag$, $assoc$]
-      <dd> represents a failure of a type indicated by $tag$, with details \
-           given by the association $assoc$.
-    </dl>
-    """
-
-    summary_text = "a failure at the level of the interpreter"
-
-
 # TODO: seems to want to produces a fancy box for failure.
 #    rules = {'Failure /: MakeBoxes[Failure[tag_, assoc_Association], StandardForm]' :
 # 		'With[{msg = assoc["MessageTemplate"], msgParam = assoc["MessageParameters"], type = assoc["Type"]}, ToBoxes @ Interpretation["Failure" @ Panel @ Grid[{{Style["\[WarningSign]", "Message", FontSize -> 35], Style["Message:", FontColor->GrayLevel[0.5]], ToString[StringForm[msg, Sequence @@ msgParam], StandardForm]}, {SpanFromAbove, Style["Tag:", FontColor->GrayLevel[0.5]], ToString[tag, StandardForm]},{SpanFromAbove,Style["Type:", FontColor->GrayLevel[0.5]],ToString[type, StandardForm]}},Alignment -> {Left, Top}], Failure[tag, assoc]] /; msg =!= Missing["KeyAbsent", "MessageTemplate"] && msgParam =!= Missing["KeyAbsent", "MessageParameters"] && msgParam =!= Missing["KeyAbsent", "Type"]]',
