+++ mathics/builtin/messages.py
@@ -6,7 +6,7 @@ Message-related functions.
 import typing
 from typing import Any
 
-from mathics.builtin.base import BinaryOperator, Builtin
+from mathics.builtin.base import BinaryOperator, Builtin, Predefined
 from mathics.core.atoms import String
 from mathics.core.attributes import A_HOLD_ALL, A_HOLD_FIRST, A_PROTECTED
 from mathics.core.evaluation import Evaluation, Message as EvaluationMessage
@@ -16,49 +16,18 @@ from mathics.core.symbols import Symbol, SymbolNull
 from mathics.core.systemsymbols import SymbolMessageName, SymbolQuiet
 
 
-class Message(Builtin):
+class Aborted(Predefined):
     """
-    <url>
-    :WMA link:
-    https://reference.wolfram.com/language/ref/Message.html</url>
+    <url>:WMA link:https://reference.wolfram.com/language/ref/Aborted.html</url>
 
     <dl>
-      <dt>'Message[$symbol$::$msg$, $expr1$, $expr2$, ...]'
-      <dd>displays the specified message, replacing placeholders in
-        the message text with the corresponding expressions.
+    <dt>'$Aborted'
+        <dd>is returned by a calculation that has been aborted.
     </dl>
-
-    >> a::b = "Hello world!"
-     = Hello world!
-    >> Message[a::b]
-     : Hello world!
-    >> a::c := "Hello `1`, Mr 00`2`!"
-    >> Message[a::c, "you", 3 + 4]
-     : Hello you, Mr 007!
     """
 
-    attributes = A_HOLD_FIRST | A_PROTECTED
-
-    messages = {
-        "name": "Message name `1` is not of the form symbol::name or symbol::name::language."
-    }
-    summary_text = "display a message"
-
-    def eval(self, symbol: Symbol, tag: String, params, evaluation: Evaluation):
-        "Message[MessageName[symbol_Symbol, tag_String], params___]"
-
-        params = params.get_sequence()
-        evaluation.message(symbol.name, tag.value, *params)
-        return SymbolNull
-
-
-def check_message(expr) -> bool:
-    "checks if an expression is a valid message"
-    if expr.has_form("MessageName", 2):
-        symbol, tag = expr.elements
-        if symbol.get_name() and tag.get_string_value():
-            return True
-    return False
+    summary_text = "return value for aborted evaluations"
+    name = "$Aborted"
 
 
 class Check(Builtin):
@@ -198,6 +167,326 @@ class Check(Builtin):
         return failexpr if display_fail_expr is True else result
 
 
+class Failed(Predefined):
+    """
+    <url>:WMA link:https://reference.wolfram.com/language/ref/$Failed.html</url>
+    <dl>
+    <dt>'$Failed'
+        <dd>is returned by some functions in the event of an error.
+    </dl>
+
+    #> Get["nonexistent_file.m"]
+     : Cannot open nonexistent_file.m.
+     = $Failed
+    """
+
+    summary_text = "retrieved result for failed evaluations"
+    name = "$Failed"
+
+
+class Failure(Builtin):
+    """
+    <url>:WMA link:https://reference.wolfram.com/language/ref/Failure.html</url>
+
+    <dl>
+      <dt>Failure[$tag$, $assoc$]
+      <dd> represents a failure of a type indicated by $tag$, with details \
+           given by the association $assoc$.
+    </dl>
+    """
+
+    summary_text = "a failure at the level of the interpreter"
+
+
+class General(Builtin):
+    """
+    <url>
+    :WMA link:
+    https://reference.wolfram.com/language/ref/General.html</url>
+
+    <dl>
+      <dt>'General'
+      <dd>is a symbol to which all general-purpose messages are assigned.
+    </dl>
+
+    >> General::argr
+     = `1` called with 1 argument; `2` arguments are expected.
+    >> Message[Rule::argr, Rule, 2]
+     : Rule called with 1 argument; 2 arguments are expected.
+    """
+
+    messages = {
+        "argb": (
+            "`1` called with `2` arguments; "
+            "between `3` and `4` arguments are expected."
+        ),
+        "argct": "`1` called with `2` arguments.",
+        "argctu": "`1` called with 1 argument.",
+        "argr": "`1` called with 1 argument; `2` arguments are expected.",
+        "argrx": "`1` called with `2` arguments; `3` arguments are expected.",
+        "argx": "`1` called with `2` arguments; 1 argument is expected.",
+        "argt": (
+            "`1` called with `2` arguments; " "`3` or `4` arguments are expected."
+        ),
+        "argtu": ("`1` called with 1 argument; `2` or `3` arguments are expected."),
+        "base": "Requested base `1` in `2` should be between 2 and `3`.",
+        "boxfmt": "`1` is not a box formatting type.",
+        "charcode": "The character encoding `1` is not supported. Use $CharacterEncodings to list supported encodings.",
+        "color": "`1` is not a valid color or gray-level specification.",
+        "cxt": "`1` is not a valid context name.",
+        "divz": "The argument `1` should be nonzero.",
+        "digit": "Digit at position `1` in `2` is too large to be used in base `3`.",
+        "exact": "Argument `1` is not an exact number.",
+        "fnsym": (
+            "First argument in `1` is not a symbol " "or a string naming a symbol."
+        ),
+        "heads": "Heads `1` and `2` are expected to be the same.",
+        "ilsnn": (
+            "Single or list of non-negative integers expected at " "position `1`."
+        ),
+        "indet": "Indeterminate expression `1` encountered.",
+        "innf": "Non-negative integer or Infinity expected at position `1`.",
+        "int": "Integer expected.",
+        "intp": "Positive integer expected.",
+        "intnn": "Non-negative integer expected.",
+        "iterb": "Iterator does not have appropriate bounds.",
+        "ivar": "`1` is not a valid variable.",
+        "level": ("Level specification `1` is not of the form n, " "{n}, or {m, n}."),
+        "locked": "Symbol `1` is locked.",
+        "matsq": "Argument `1` is not a non-empty square matrix.",
+        "newpkg": "In WL, there is a new package for this.",
+        "noopen": "Cannot open `1`.",
+        "nord": "Invalid comparison with `1` attempted.",
+        "normal": "Nonatomic expression expected.",
+        "noval": ("Symbol `1` in part assignment does not have an immediate value."),
+        "obspkg": "In WL, this package is obsolete.",
+        "openx": "`1` is not open.",
+        "optb": "Optional object `1` in `2` is not a single blank.",
+        "ovfl": "Overflow occurred in computation.",
+        "partd": "Part specification is longer than depth of object.",
+        "partw": "Part `1` of `2` does not exist.",
+        "plld": "Endpoints in `1` must be distinct machine-size real numbers.",
+        "plln": "Limiting value `1` in `2` is not a machine-size real number.",
+        "pspec": (
+            "Part specification `1` is neither an integer nor " "a list of integer."
+        ),
+        "seqs": "Sequence specification expected, but got `1`.",
+        "setp": "Part assignment to `1` could not be made",
+        "setps": "`1` in the part assignment is not a symbol.",
+        "span": "`1` is not a valid Span specification.",
+        "ssym": "`1` is not a symbol or a string.",
+        "stream": "`1` is not string, InputStream[], or OutputStream[]",
+        "string": "String expected.",
+        "sym": "Argument `1` at position `2` is expected to be a symbol.",
+        "tag": "Rule for `1` can only be attached to `2`.",
+        "take": "Cannot take positions `1` through `2` in `3`.",
+        "ucdec": "An invalid unicode sequence was encountered and ignored.",
+        "vrule": (
+            "Cannot set `1` to `2`, " "which is not a valid list of replacement rules."
+        ),
+        "write": "Tag `1` in `2` is Protected.",
+        "wrsym": "Symbol `1` is Protected.",
+        # TODO: someone please explain why these are different...
+        # Self-defined messages
+        "rep": "`1` is not a valid replacement rule.",
+        "options": "`1` is not a valid list of option rules.",
+        "timeout": "Timeout reached.",
+        "syntax": "`1`",
+        "invalidargs": "Invalid arguments.",
+        "notboxes": "`1` is not a valid box structure.",
+        "pyimport": '`1`[] is not available. Python module "`2`" is not installed.',
+    }
+    summary_text = "general-purpose messages"
+
+
+class Message(Builtin):
+    """
+    <url>
+    :WMA link:
+    https://reference.wolfram.com/language/ref/Message.html</url>
+
+    <dl>
+      <dt>'Message[$symbol$::$msg$, $expr1$, $expr2$, ...]'
+      <dd>displays the specified message, replacing placeholders in
+        the message text with the corresponding expressions.
+    </dl>
+
+    >> a::b = "Hello world!"
+     = Hello world!
+    >> Message[a::b]
+     : Hello world!
+    >> a::c := "Hello `1`, Mr 00`2`!"
+    >> Message[a::c, "you", 3 + 4]
+     : Hello you, Mr 007!
+    """
+
+    attributes = A_HOLD_FIRST | A_PROTECTED
+
+    messages = {
+        "name": "Message name `1` is not of the form symbol::name or symbol::name::language."
+    }
+    summary_text = "display a message"
+
+    def eval(self, symbol: Symbol, tag: String, params, evaluation: Evaluation):
+        "Message[MessageName[symbol_Symbol, tag_String], params___]"
+
+        params = params.get_sequence()
+        evaluation.message(symbol.name, tag.value, *params)
+        return SymbolNull
+
+
+def check_message(expr) -> bool:
+    "checks if an expression is a valid message"
+    if expr.has_form("MessageName", 2):
+        symbol, tag = expr.elements
+        if symbol.get_name() and tag.get_string_value():
+            return True
+    return False
+
+
+class MessageName(BinaryOperator):
+    """
+    <url>:WMA link:https://reference.wolfram.com/language/ref/MessageName.html</url>
+
+    <dl>
+      <dt>'MessageName[$symbol$, $tag$]'
+      <dt>'$symbol$::$tag$'
+      <dd>identifies a message.
+    </dl>
+
+    'MessageName' is the head of message IDs of the form 'symbol::tag'.
+    >> FullForm[a::b]
+     = MessageName[a, "b"]
+
+    The second parameter 'tag' is interpreted as a string.
+    >> FullForm[a::"b"]
+     = MessageName[a, "b"]
+    """
+
+    attributes = A_HOLD_FIRST | A_PROTECTED
+    default_formats = False
+    formats: typing.Dict[str, Any] = {}
+    messages = {"messg": "Message cannot be set to `1`. It must be set to a string."}
+    summary_text = "message identifyier"
+    operator = "::"
+    precedence = 750
+    rules = {
+        "MakeBoxes[MessageName[symbol_Symbol, tag_String], "
+        "f:StandardForm|TraditionalForm|OutputForm]": (
+            'RowBox[{MakeBoxes[symbol, f], "::", MakeBoxes[tag, f]}]'
+        ),
+        "MakeBoxes[MessageName[symbol_Symbol, tag_String], InputForm]": (
+            'RowBox[{MakeBoxes[symbol, InputForm], "::", tag}]'
+        ),
+    }
+
+    def eval(self, symbol: Symbol, tag: String, evaluation: Evaluation):
+        "MessageName[symbol_Symbol, tag_String]"
+
+        pattern = Expression(SymbolMessageName, symbol, tag)
+        return evaluation.definitions.get_value(
+            symbol.get_name(), "System`Messages", pattern, evaluation
+        )
+
+
+class Off(Builtin):
+    """
+    <url>:WMA link:https://reference.wolfram.com/language/ref/Off.html</url>
+
+    <dl>
+      <dt>'Off[$symbol$::$tag$]'
+      <dd>turns a message off so it is no longer printed.
+    </dl>
+
+    >> Off[Power::infy]
+    >> 1 / 0
+     = ComplexInfinity
+
+    >> Off[Power::indet, Syntax::com]
+    >> {0 ^ 0,}
+     = {Indeterminate, Null}
+
+    #> Off[1]
+     :  Message name 1 is not of the form symbol::name or symbol::name::language.
+    #> Off[Message::name, 1]
+
+    #> On[Power::infy, Power::indet, Syntax::com]
+    """
+
+    attributes = A_HOLD_ALL | A_PROTECTED
+    summary_text = "turn off a message for printing"
+
+    def eval(self, expr, evaluation: Evaluation):
+        "Off[expr___]"
+
+        seq = expr.get_sequence()
+        quiet_messages = set(evaluation.get_quiet_messages())
+
+        if not seq:
+            # TODO Off[s::trace] for all symbols
+            return
+
+        for e in seq:
+            if isinstance(e, Symbol):
+                quiet_messages.add(Expression(SymbolMessageName, e, String("trace")))
+            elif check_message(e):
+                quiet_messages.add(e)
+            else:
+                evaluation.message("Message", "name", e)
+            evaluation.set_quiet_messages(quiet_messages)
+
+        return SymbolNull
+
+
+class On(Builtin):
+    """
+    <url>:WMA link:https://reference.wolfram.com/language/ref/On.html</url>
+
+    <dl>
+      <dt>'On[$symbol$::$tag$]'
+      <dd>turns a message on for printing.
+    </dl>
+
+    >> Off[Power::infy]
+    >> 1 / 0
+     = ComplexInfinity
+    >> On[Power::infy]
+    >> 1 / 0
+     : Infinite expression 1 / 0 encountered.
+     = ComplexInfinity
+    """
+
+    # TODO
+    """
+    #> On[f::x]
+     : Message f::x not found.
+    """
+    attributes = A_HOLD_ALL | A_PROTECTED
+    summary_text = "turn on a message for printing"
+
+    def eval(self, expr, evaluation: Evaluation):
+        "On[expr___]"
+
+        seq = expr.get_sequence()
+        quiet_messages = set(evaluation.get_quiet_messages())
+
+        if not seq:
+            # TODO On[s::trace] for all symbols
+            return
+
+        for e in seq:
+            if isinstance(e, Symbol):
+                quiet_messages.discard(
+                    Expression(SymbolMessageName, e, String("trace"))
+                )
+            elif check_message(e):
+                quiet_messages.discard(e)
+            else:
+                evaluation.message("Message", "name", e)
+            evaluation.set_quiet_messages(quiet_messages)
+        return SymbolNull
+
+
 class Quiet(Builtin):
     """
     <url>:WMA link:https://reference.wolfram.com/language/ref/Quiet.html</url>
@@ -328,149 +617,6 @@ class Quiet(Builtin):
             evaluation.set_quiet_messages(old_quiet_messages)
 
 
-class Off(Builtin):
-    """
-    <url>:WMA link:https://reference.wolfram.com/language/ref/Off.html</url>
-
-    <dl>
-      <dt>'Off[$symbol$::$tag$]'
-      <dd>turns a message off so it is no longer printed.
-    </dl>
-
-    >> Off[Power::infy]
-    >> 1 / 0
-     = ComplexInfinity
-
-    >> Off[Power::indet, Syntax::com]
-    >> {0 ^ 0,}
-     = {Indeterminate, Null}
-
-    #> Off[1]
-     :  Message name 1 is not of the form symbol::name or symbol::name::language.
-    #> Off[Message::name, 1]
-
-    #> On[Power::infy, Power::indet, Syntax::com]
-    """
-
-    attributes = A_HOLD_ALL | A_PROTECTED
-    summary_text = "turn off a message for printing"
-
-    def eval(self, expr, evaluation: Evaluation):
-        "Off[expr___]"
-
-        seq = expr.get_sequence()
-        quiet_messages = set(evaluation.get_quiet_messages())
-
-        if not seq:
-            # TODO Off[s::trace] for all symbols
-            return
-
-        for e in seq:
-            if isinstance(e, Symbol):
-                quiet_messages.add(Expression(SymbolMessageName, e, String("trace")))
-            elif check_message(e):
-                quiet_messages.add(e)
-            else:
-                evaluation.message("Message", "name", e)
-            evaluation.set_quiet_messages(quiet_messages)
-
-        return SymbolNull
-
-
-class On(Builtin):
-    """
-    <url>:WMA link:https://reference.wolfram.com/language/ref/On.html</url>
-
-    <dl>
-      <dt>'On[$symbol$::$tag$]'
-      <dd>turns a message on for printing.
-    </dl>
-
-    >> Off[Power::infy]
-    >> 1 / 0
-     = ComplexInfinity
-    >> On[Power::infy]
-    >> 1 / 0
-     : Infinite expression 1 / 0 encountered.
-     = ComplexInfinity
-    """
-
-    # TODO
-    """
-    #> On[f::x]
-     : Message f::x not found.
-    """
-    attributes = A_HOLD_ALL | A_PROTECTED
-    summary_text = "turn on a message for printing"
-
-    def eval(self, expr, evaluation: Evaluation):
-        "On[expr___]"
-
-        seq = expr.get_sequence()
-        quiet_messages = set(evaluation.get_quiet_messages())
-
-        if not seq:
-            # TODO On[s::trace] for all symbols
-            return
-
-        for e in seq:
-            if isinstance(e, Symbol):
-                quiet_messages.discard(
-                    Expression(SymbolMessageName, e, String("trace"))
-                )
-            elif check_message(e):
-                quiet_messages.discard(e)
-            else:
-                evaluation.message("Message", "name", e)
-            evaluation.set_quiet_messages(quiet_messages)
-        return SymbolNull
-
-
-class MessageName(BinaryOperator):
-    """
-    <url>:WMA link:https://reference.wolfram.com/language/ref/MessageName.html</url>
-
-    <dl>
-      <dt>'MessageName[$symbol$, $tag$]'
-      <dt>'$symbol$::$tag$'
-      <dd>identifies a message.
-    </dl>
-
-    'MessageName' is the head of message IDs of the form 'symbol::tag'.
-    >> FullForm[a::b]
-     = MessageName[a, "b"]
-
-    The second parameter 'tag' is interpreted as a string.
-    >> FullForm[a::"b"]
-     = MessageName[a, "b"]
-    """
-
-    attributes = A_HOLD_FIRST | A_PROTECTED
-    default_formats = False
-    formats: typing.Dict[str, Any] = {}
-    messages = {"messg": "Message cannot be set to `1`. It must be set to a string."}
-    summary_text = "message identifyier"
-    operator = "::"
-    precedence = 750
-    rules = {
-        "MakeBoxes[MessageName[symbol_Symbol, tag_String], "
-        "f:StandardForm|TraditionalForm|OutputForm]": (
-            'RowBox[{MakeBoxes[symbol, f], "::", MakeBoxes[tag, f]}]'
-        ),
-        "MakeBoxes[MessageName[symbol_Symbol, tag_String], InputForm]": (
-            'RowBox[{MakeBoxes[symbol, InputForm], "::", tag}]'
-        ),
-    }
-
-    def eval(self, symbol: Symbol, tag: String, evaluation: Evaluation):
-        "MessageName[symbol_Symbol, tag_String]"
-
-        pattern = Expression(SymbolMessageName, symbol, tag)
-        return evaluation.definitions.get_value(
-            symbol.get_name(), "System`Messages", pattern, evaluation
-        )
-
-
 class Syntax(Builtin):
     r"""
     <url>:WMA link:https://reference.wolfram.com/language/ref/Syntax.html</url>
@@ -575,104 +721,3 @@ class Syntax(Builtin):
         "com": "Warning: comma encountered with no adjacent expression. The expression will be treated as Null (line `4` of `5`).",
     }
     summary_text = "syntax messages"
-
-
-class General(Builtin):
-    """
-    <url>
-    :WMA link:
-    https://reference.wolfram.com/language/ref/General.html</url>
-
-    <dl>
-      <dt>'General'
-      <dd>is a symbol to which all general-purpose messages are assigned.
-    </dl>
-
-    >> General::argr
-     = `1` called with 1 argument; `2` arguments are expected.
-    >> Message[Rule::argr, Rule, 2]
-     : Rule called with 1 argument; 2 arguments are expected.
-    """
-
-    messages = {
-        "argb": (
-            "`1` called with `2` arguments; "
-            "between `3` and `4` arguments are expected."
-        ),
-        "argct": "`1` called with `2` arguments.",
-        "argctu": "`1` called with 1 argument.",
-        "argr": "`1` called with 1 argument; `2` arguments are expected.",
-        "argrx": "`1` called with `2` arguments; `3` arguments are expected.",
-        "argx": "`1` called with `2` arguments; 1 argument is expected.",
-        "argt": (
-            "`1` called with `2` arguments; " "`3` or `4` arguments are expected."
-        ),
-        "argtu": ("`1` called with 1 argument; `2` or `3` arguments are expected."),
-        "base": "Requested base `1` in `2` should be between 2 and `3`.",
-        "boxfmt": "`1` is not a box formatting type.",
-        "charcode": "The character encoding `1` is not supported. Use $CharacterEncodings to list supported encodings.",
-        "color": "`1` is not a valid color or gray-level specification.",
-        "cxt": "`1` is not a valid context name.",
-        "divz": "The argument `1` should be nonzero.",
-        "digit": "Digit at position `1` in `2` is too large to be used in base `3`.",
-        "exact": "Argument `1` is not an exact number.",
-        "fnsym": (
-            "First argument in `1` is not a symbol " "or a string naming a symbol."
-        ),
-        "heads": "Heads `1` and `2` are expected to be the same.",
-        "ilsnn": (
-            "Single or list of non-negative integers expected at " "position `1`."
-        ),
-        "indet": "Indeterminate expression `1` encountered.",
-        "innf": "Non-negative integer or Infinity expected at position `1`.",
-        "int": "Integer expected.",
-        "intp": "Positive integer expected.",
-        "intnn": "Non-negative integer expected.",
-        "iterb": "Iterator does not have appropriate bounds.",
-        "ivar": "`1` is not a valid variable.",
-        "level": ("Level specification `1` is not of the form n, " "{n}, or {m, n}."),
-        "locked": "Symbol `1` is locked.",
-        "matsq": "Argument `1` is not a non-empty square matrix.",
-        "newpkg": "In WL, there is a new package for this.",
-        "noopen": "Cannot open `1`.",
-        "nord": "Invalid comparison with `1` attempted.",
-        "normal": "Nonatomic expression expected.",
-        "noval": ("Symbol `1` in part assignment does not have an immediate value."),
-        "obspkg": "In WL, this package is obsolete.",
-        "openx": "`1` is not open.",
-        "optb": "Optional object `1` in `2` is not a single blank.",
-        "ovfl": "Overflow occurred in computation.",
-        "partd": "Part specification is longer than depth of object.",
-        "partw": "Part `1` of `2` does not exist.",
-        "plld": "Endpoints in `1` must be distinct machine-size real numbers.",
-        "plln": "Limiting value `1` in `2` is not a machine-size real number.",
-        "pspec": (
-            "Part specification `1` is neither an integer nor " "a list of integer."
-        ),
-        "seqs": "Sequence specification expected, but got `1`.",
-        "setp": "Part assignment to `1` could not be made",
-        "setps": "`1` in the part assignment is not a symbol.",
-        "span": "`1` is not a valid Span specification.",
-        "ssym": "`1` is not a symbol or a string.",
-        "stream": "`1` is not string, InputStream[], or OutputStream[]",
-        "string": "String expected.",
-        "sym": "Argument `1` at position `2` is expected to be a symbol.",
-        "tag": "Rule for `1` can only be attached to `2`.",
-        "take": "Cannot take positions `1` through `2` in `3`.",
-        "ucdec": "An invalid unicode sequence was encountered and ignored.",
-        "vrule": (
-            "Cannot set `1` to `2`, " "which is not a valid list of replacement rules."
-        ),
-        "write": "Tag `1` in `2` is Protected.",
-        "wrsym": "Symbol `1` is Protected.",
-        # TODO: someone please explain why these are different...
-        # Self-defined messages
-        "rep": "`1` is not a valid replacement rule.",
-        "options": "`1` is not a valid list of option rules.",
-        "timeout": "Timeout reached.",
-        "syntax": "`1`",
-        "invalidargs": "Invalid arguments.",
-        "notboxes": "`1` is not a valid box structure.",
-        "pyimport": '`1`[] is not available. Python module "`2`" is not installed.',
-    }
-    summary_text = "general-purpose messages"
